fix type

+++ src/diffusers/models/autoencoder_kl.py
@@ -78,7 +78,7 @@ class AutoencoderKL(ModelMixin, ConfigMixin):
         latent_channels: int = 4,
         norm_num_groups: int = 32,
         sample_size: int = 32,
-        scaling_factor: int = 0.18215,
+        scaling_factor: float = 0.18215,
     ):
         super().__init__()
