+++ src/diffusers/models/vae_flax.py
@@ -774,7 +774,7 @@ class FlaxAutoencoderKL(nn.Module, FlaxModelMixin, ConfigMixin):
     latent_channels: int = 4
     norm_num_groups: int = 32
     sample_size: int = 32
-    scaling_factor: int = (0.18215,)
+    scaling_factor: float = 0.18215
     dtype: jnp.dtype = jnp.float32
 
     def setup(self):
